+++ app/src/main/java/com/sekusarisu/mdpings/vpings/data/networking/dto/DashboardServersResponsesDto.kt
@@ -3,6 +3,7 @@ package com.sekusarisu.mdpings.vpings.data.networking.dto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MonitorsResponsesDto(
-    val data: List<MonitorDto>? = emptyList<MonitorDto>()
+data class DashboardServersResponsesDto(
+    val data: List<DashboardServerDto> = emptyList<DashboardServerDto>()
 )
+