+++ app/src/main/java/com/sekusarisu/mdpings/vpings/data/networking/dto/LoginDataDto.kt
@@ -3,6 +3,7 @@ package com.sekusarisu.mdpings.vpings.data.networking.dto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MonitorsResponsesDto(
-    val data: List<MonitorDto>? = emptyList<MonitorDto>()
+data class LoginDataDto(
+    val token: String,
+    val expire: String
 )