+++ app/src/main/java/com/sekusarisu/mdpings/vpings/data/networking/dto/LoginResponsesDto.kt
@@ -3,6 +3,6 @@ package com.sekusarisu.mdpings.vpings.data.networking.dto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MonitorsResponsesDto(
-    val data: List<MonitorDto>? = emptyList<MonitorDto>()
+data class LoginResponsesDto(
+    val data: LoginDataDto? = LoginDataDto("", "")
 )