+++ app/src/main/java/com/sekusarisu/mdpings/vpings/presentation/server_detail/ServerDetailScreen.kt
@@ -47,61 +47,53 @@ import com.sekusarisu.mdpings.ui.theme.MDPingsTheme
 import com.sekusarisu.mdpings.vpings.domain.AppSettings
 import com.sekusarisu.mdpings.vpings.presentation.app_settings.AppSettingsState
 import com.sekusarisu.mdpings.vpings.presentation.models.ServerUi
+import com.sekusarisu.mdpings.vpings.presentation.models.WSServerUi
 import com.sekusarisu.mdpings.vpings.presentation.server_detail.components.InstanceInfo
 import com.sekusarisu.mdpings.vpings.presentation.server_detail.components.NetworkMonitor
 import com.sekusarisu.mdpings.vpings.presentation.server_detail.components.OfflineCard
 import com.sekusarisu.mdpings.vpings.presentation.server_detail.components.PktLostAndAvgLatencyCard
 import com.sekusarisu.mdpings.vpings.presentation.server_detail.components.mockIpAPIUi
 import com.sekusarisu.mdpings.vpings.presentation.server_detail.components.mockMonitors
+import com.sekusarisu.mdpings.vpings.presentation.server_list.ServerListState
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.LoadAndUptime
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.NetworkIO
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.NetworkTransfer
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.OnlineStatusIndicator
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.Status
+import com.sekusarisu.mdpings.vpings.presentation.server_list.components.previewListServers
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.previewServerUi0
 import com.sekusarisu.mdpings.vpings.presentation.server_list.components.previewServerUi1
+import com.sekusarisu.mdpings.vpings.presentation.server_list.components.previewWSServerUi0
+import com.sekusarisu.mdpings.vpings.presentation.server_list.components.previewWSServerUi1
 import kotlinx.coroutines.delay
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ServerDetailScreen(
     state: ServerDetailState,
+    serverListState: ServerListState,
     appSettingsState: AppSettingsState,
-    selectedServerUi: ServerUi,
+    selectedServerUi: WSServerUi,
     onAction: (ServerDetailAction) -> Unit,
     modifier: Modifier = Modifier
 ) {
     val lifecycleOwner = LocalLifecycleOwner.current
 
+    // Load Info & Monitors
     LaunchedEffect(selectedServerUi.id) {
-        val apiURL = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].apiUrl
-        val apiTOKEN = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].apiToken
+        val baseUrl = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].baseUrl
+        val username = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].username
+        val password = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].password
         val interval = appSettingsState.appSettings.refreshInterval
         onAction(
             ServerDetailAction.OnLoadInfoAndMonitors(
                 serverUi = selectedServerUi,
                 monitorsTimeSlice = state.monitorsTimeSlice,
-                apiURL = apiURL,
-                apiTOKEN = apiTOKEN,
+                apiURL = baseUrl,
+                apiTOKEN = username,
                 interval = interval
             )
         )
-    }
-    LaunchedEffect(selectedServerUi.id) {
-        val apiURL = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].apiUrl
-        val apiTOKEN = appSettingsState.appSettings.instances[appSettingsState.appSettings.activeInstance].apiToken
-        val interval = appSettingsState.appSettings.refreshInterval
-        while (lifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
-            onAction(
-                ServerDetailAction.OnLoadSingleServer(
-                    serverUi = selectedServerUi,
-                    apiURL = apiURL,
-                    apiTOKEN = apiTOKEN,
-                    interval = interval
-                )
-            )
-            delay(interval.toLong())
-        }
     }
     DisposableEffect(lifecycleOwner) {
         onDispose{
@@ -112,7 +104,8 @@ fun ServerDetailScreen(
     }
 
     AnimatedVisibility(
-        visible = (state.serverUi == null || state.ipAPIUi == null),
+        visible = (state.wsServerUi == null || state.ipAPIUi == null),
+//        visible = (state.wsServerUi == null),
         enter = fadeIn(),
         exit = fadeOut()
     ) {
@@ -129,7 +122,8 @@ fun ServerDetailScreen(
     }
 
     AnimatedVisibility(
-        visible = (state.serverUi != null && state.ipAPIUi != null),
+        visible = (state.wsServerUi != null && state.ipAPIUi != null),
+//        visible = (state.wsServerUi != null),
         enter = slideInVertically(
             spring(
                 stiffness = Spring.StiffnessLow,
@@ -148,12 +142,12 @@ fun ServerDetailScreen(
                 .verticalScroll(rememberScrollState())
                 .padding(horizontal = 8.dp, vertical = 4.dp)
         ) {
-            if (!state.serverUi!!.isOnline) {
+            if (!selectedServerUi.isOnline) {
                 OfflineCard()
                 Spacer(Modifier.height(8.dp))
             }
             InstanceInfo(
-                serverUi = state.serverUi,
+                serverUi = state.wsServerUi!!,
                 ipAPIUi = state.ipAPIUi!!,
                 modifier = modifier
             )
@@ -162,7 +156,8 @@ fun ServerDetailScreen(
                 appSettingsState = appSettingsState,
                 state = state,
                 onAction = onAction,
-                modifier = modifier
+                modifier = modifier,
+                serverId = selectedServerUi.id
             )
             Spacer(Modifier.height(8.dp))
             PktLostAndAvgLatencyCard(
@@ -172,7 +167,8 @@ fun ServerDetailScreen(
             )
             Spacer(Modifier.height(8.dp))
             ServerStatus(
-                serverUi = state.serverUi,
+                state = serverListState,
+                selectedServerUi = selectedServerUi,
                 modifier = modifier
             )
         }
@@ -181,9 +177,13 @@ fun ServerDetailScreen(
 
 @Composable
 fun ServerStatus(
-    serverUi: ServerUi,
+    state: ServerListState,
+    selectedServerUi: WSServerUi,
     modifier: Modifier = Modifier
 ) {
+
+    val serverUi = state.wsServers.first { server -> server.id == selectedServerUi.id }
+
     Card(
         modifier = Modifier.wrapContentHeight(),
         shape = ShapeDefaults.Medium,
@@ -239,16 +239,25 @@ fun ServerStatus(
 fun ServerDetailScreenPreviewOnline() {
     MDPingsTheme {
         ServerDetailScreen(
-            selectedServerUi = previewServerUi0,
+            selectedServerUi = previewWSServerUi0,
             modifier = Modifier,
             state = ServerDetailState(
                 isLoading = false,
                 serverUi = previewServerUi0,
+                wsServerUi = previewWSServerUi0,
                 ipAPIUi = mockIpAPIUi,
                 monitors = mockMonitors
             ),
             onAction = {},
-            appSettingsState = AppSettingsState(AppSettings())
+            appSettingsState = AppSettingsState(AppSettings()),
+            serverListState = ServerListState(
+                isLoading = false,
+                selectedServer = previewWSServerUi1,
+                servers = previewListServers,
+                wsServers = listOf(previewWSServerUi0, previewWSServerUi1),
+                ipAPIUi = mockIpAPIUi,
+                monitors = mockMonitors
+            )
         )
     }
 }
@@ -258,16 +267,25 @@ fun ServerDetailScreenPreviewOnline() {
 fun ServerDetailScreenPreviewOffline() {
     MDPingsTheme {
         ServerDetailScreen(
-            selectedServerUi = previewServerUi1,
+            selectedServerUi = previewWSServerUi1,
             modifier = Modifier,
             state = ServerDetailState(
                 isLoading = false,
                 serverUi = previewServerUi1,
+                wsServerUi = previewWSServerUi1,
                 ipAPIUi = mockIpAPIUi,
                 monitors = mockMonitors
             ),
             onAction = {},
-            appSettingsState = AppSettingsState(AppSettings())
+            appSettingsState = AppSettingsState(AppSettings()),
+            serverListState = ServerListState(
+                isLoading = false,
+                selectedServer = previewWSServerUi1,
+                servers = previewListServers,
+                wsServers = listOf(previewWSServerUi0, previewWSServerUi1),
+                ipAPIUi = mockIpAPIUi,
+                monitors = mockMonitors
+            )
         )
     }
 }