+++ app/src/main/java/com/sekusarisu/mdpings/vpings/presentation/user_login/LoginViewModel.kt
@@ -2,16 +2,20 @@ package com.sekusarisu.mdpings.vpings.presentation.user_login
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.sekusarisu.mdpings.core.domain.util.Error
+import com.sekusarisu.mdpings.core.domain.util.NetworkError
+import com.sekusarisu.mdpings.core.domain.util.Result
 import com.sekusarisu.mdpings.core.domain.util.onError
 import com.sekusarisu.mdpings.core.domain.util.onSuccess
 import com.sekusarisu.mdpings.vpings.domain.AppSettingsDataSource
+import com.sekusarisu.mdpings.vpings.domain.LoginData
 import com.sekusarisu.mdpings.vpings.domain.ServerDataSource
-import com.sekusarisu.mdpings.vpings.presentation.models.toServerUi
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import kotlin.fold
 
 class LoginViewModel(
     private val serverDataSource: ServerDataSource,
@@ -31,22 +35,20 @@ class LoginViewModel(
             is LoginAction.OnInitLoadInstances -> {
                 getInstances()
             }
-            is LoginAction.OnTestClick -> {
-                testConnection(
-                    apiUrl = action.apiURL,
-                    token = action.apiTOKEN
-                )
-            }
             is LoginAction.OnCredentialsChange -> {
                 _state.update { it.copy(
-                    servers = emptyList()
+                    servers = emptyList(),
+                    isTestSucceed = false
                 ) }
             }
+            is LoginAction.OnTestClick -> {
+                testConnection(action.baseUrl, action.username, action.password)
+            }
             is LoginAction.OnSaveClicked -> {
-                saveInstance(action.name, action.apiURL, action.apiTOKEN)
+                saveInstance(action.name, action.baseUrl, action.username, action.password)
             }
             is LoginAction.OnEditSaveClicked -> {
-                editInstance(action.index, action.name, action.apiURL, action.apiTOKEN)
+                editInstance(action.index, action.name, action.baseUrl, action.username, action.password)
             }
             is LoginAction.OnDeleteClick -> {
                 deleteInstance(action.index)
@@ -54,25 +56,56 @@ class LoginViewModel(
         }
     }
 
-    fun testConnection(apiUrl: String, token: String) {
+    fun testConnection(baseUrl: String, username: String, password: String) {
         viewModelScope.launch{
             _state.update { it.copy(
                 isLoading = true,
-                servers = emptyList()
+                isTestSucceed = false,
             ) }
-        serverDataSource
-            .getServers(apiUrl, token)
-            .onSuccess { servers ->
-                _state.update { it.copy(
-                    isLoading = false,
-                    servers = servers.map { it.toServerUi() }
-                ) }
-            }
-            .onError { error ->
-                _state.update { it.copy(isLoading = false) }
-                _events.send(LoginEvent.Error(error))
-            }
+            serverDataSource
+                .getLoginData(baseUrl, username, password)
+                .onSuccess { servers ->
+                    if (servers.token.isNotEmpty()) {
+                        _state.update { it.copy(
+                            isLoading = false,
+                            isTestSucceed = true
+                        ) }
+                    }
+                }
+                .onError { error ->
+                    _state.update { it.copy(isLoading = false) }
+                    _events.send(LoginEvent.Error(error))
+                }
         }
+    }
+
+    suspend fun testConnectionToBoolean(baseUrl: String, username: String, password: String): Boolean {
+        _state.update { it.copy(
+            isLoading = true,
+            isTestSucceed = false,
+        ) }
+        return serverDataSource
+            .getLoginData(baseUrl, username, password)
+            .let { result ->
+                when (result) {
+                    is Result.Success -> {
+                        val loginData = result.data
+                        val isSuccess = loginData.token.isNotEmpty()
+                        _state.update {
+                            it.copy(
+                                isLoading = false,
+                                isTestSucceed = isSuccess
+                            )
+                        }
+                        isSuccess
+                    }
+                    is Result.Error -> {
+                        _state.update { it.copy(isLoading = false) }
+                        _events.send(LoginEvent.Error(result.error))
+                        false
+                    }
+                }
+            }
     }
 
     fun getInstances() {
@@ -88,24 +121,24 @@ class LoginViewModel(
         }
     }
 
-    fun saveInstance(name: String, apiUrl: String, apiToken: String) {
+    fun saveInstance(name: String, baseUrl: String, username: String, password: String) {
         viewModelScope.launch{
             _state.update { it.copy(
                 isLoading = true
             ) }
-            appSettingsDataSource.putInstance(name, apiUrl, apiToken)
+            appSettingsDataSource.putInstance(name, baseUrl, username, password)
             _state.update { it.copy(
                 isLoading = false
             ) }
         }
     }
 
-    fun editInstance(index: Int, name: String, apiUrl: String, apiToken: String) {
+    fun editInstance(index: Int, name: String, baseUrl: String, username: String, password: String) {
         viewModelScope.launch{
             _state.update { it.copy(
                 isLoading = true
             ) }
-            appSettingsDataSource.editInstance(index, name, apiUrl, apiToken)
+            appSettingsDataSource.editInstance(index, name, baseUrl, username, password)
             _state.update { it.copy(
                 isLoading = false
             ) }